Move shared model overview components to lib subfolder and add SeparatedItems component
